ft: add cluster functionality
