+++ package.json
@@ -7,6 +7,7 @@
     "build": "tsc",
     "start:dev": "nodemon src/index.ts",
     "start:prod": "npm run build && node dist/index.js",
+    "start:multi": "npm run build && node dist/cluster.js",
     "test": "jest"
   },
   "keywords": [],
